Improve snake source codes
I make it more readable and understandable
